moved commands folder into cli folder
