+++ internal/cli/commands/completion.go
