+++ internal/cli/commands/example.go
