+++ internal/cli/root.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/alexchao26/oneterminal/internal/commands"
+	"github.com/alexchao26/oneterminal/internal/cli/commands"
 	"github.com/alexchao26/oneterminal/internal/monitor"
 	"github.com/alexchao26/oneterminal/internal/yaml"
 	"github.com/spf13/cobra"
